Remove create button from app analytics
